+++ packages/datasource-toolkit/src/index.ts
@@ -60,3 +60,4 @@ export { default as SortValidator } from './validation/sort';
 export { default as CollectionUtils } from './utils/collection';
 export { default as RecordUtils } from './utils/record';
 export { default as SchemaUtils } from './utils/schema';
+export { default as Deferred } from './deferred';