+++ packages/forestadmin-client/CHANGELOG.md
@@ -1,3 +1,13 @@
+## @forestadmin/forestadmin-client [1.5.10](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.5.9...@forestadmin/forestadmin-client@1.5.10) (2023-08-11)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-toolkit:** upgraded to 1.9.0
+
 # @forestadmin/forestadmin-client [1.6.0-alpha-widgets.2](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/forestadmin-client@1.6.0-alpha-widgets.1...@forestadmin/forestadmin-client@1.6.0-alpha-widgets.2) (2023-08-23)